requirements.txt
# UI
gradio[oauth]>=4.0.0

# LLM + Embeddings
openai>=1.0.0

# Vector DB
chromadb>=0.5.0

# Document parsing
PyPDF2>=3.0.0
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
requests>=2.31.0

# Audio / TTS
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0